fix bugs with choice, add printing and basic tests
choice prefers leftmost parsers if they succeed.
sort alternatives by longest length first.

this happens because the `All parser mode naively appends a <* eof
parser.
this means that when parsers like `("a" <|> "aa") <* eof`
are executed on "aa", it will match the "a" case, then fail because
the eof fails outside of the backtracking choice.
if, instead, <* eof was pushed into the end of each choice disjunction,
it would backtrack successfully.

hopefully this doesn't cause us more problems in future, as most
syntax rules wont overlap in this way (or do so in a trivial way,
e.g. the optional handling).